refactor: isolate diagonal schedule translator timing
