+++ src/tqec/compile/specs/library/generators/diagonal_schedule.py
@@ -8,16 +8,23 @@ from __future__ import annotations
 
 from typing import Literal
 
+from tqec.compile.specs.enums import SpatialArms
 from tqec.compile.specs.library.generators.fixed_bulk import FixedBulkConventionGenerator
 from tqec.plaquette.compilation.base import PlaquetteCompiler
 from tqec.plaquette.compilation.passes.scheduling import ChangeSchedulePass
 from tqec.plaquette.compilation.passes.sort_targets import SortTargetsPass
-from tqec.plaquette.rpng import RPNGDescription
-from tqec.plaquette.rpng.translators.default import DefaultRPNGTranslator
-from tqec.utils.enums import Basis, Orientation
 from tqec.plaquette.enums import PlaquetteOrientation
+from tqec.plaquette.rpng import RPNGDescription
+from tqec.plaquette.rpng.translators.diagonal import DiagonalRPNGTranslator
+from tqec.templates.qubit import (
+    QubitHorizontalBorders,
+    QubitSpatialCubeTemplate,
+    QubitTemplate,
+    QubitVerticalBorders,
+)
+from tqec.utils.enums import Basis, Orientation
+from tqec.utils.exceptions import TQECError
 from tqec.utils.frozendefaultdict import FrozenDefaultDict
-from tqec.compile.specs.enums import SpatialArms
 
 
 def create_diagonal_schedule_compiler() -> PlaquetteCompiler:
@@ -37,31 +44,32 @@ def create_diagonal_schedule_compiler() -> PlaquetteCompiler:
 
 class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
     """Generator with diagonal plaquette configurations for syndrome extraction.
-    
+
     This generator implements diagonal schedules:
     - X-basis bulk: schedules [1, 4, 3, 2] (instead of [1, 2, 3, 5])
     - Z-basis bulk: schedules [6, 4, 3, 5] (instead of [1, 2, 3, 5])
-    
+
     The diagonal schedule enables diagonal syndrome extraction patterns.
     """
-    
+
     def __init__(
         self,
         translator=None,
         compiler=None,
     ):
         """Initialize the diagonal schedule generator.
-        
+
         Args:
-            translator: RPNG translator instance. Defaults to DefaultRPNGTranslator.
+            translator: RPNG translator instance. Defaults to DiagonalRPNGTranslator.
             compiler: Plaquette compiler instance. Defaults to diagonal schedule compiler.
+
         """
         if translator is None:
-            translator = DefaultRPNGTranslator()
+            translator = DiagonalRPNGTranslator()
         if compiler is None:
             compiler = create_diagonal_schedule_compiler()
         super().__init__(translator, compiler)
-    
+
     def get_diagonal_bulk_rpng_descriptions(
         self,
         reset: Basis | None = None,
@@ -69,23 +77,24 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
         reset_and_measured_indices: tuple[Literal[0, 1, 2, 3], ...] = (0, 1, 2, 3),
     ) -> dict[Basis, dict[Orientation, RPNGDescription]]:
         """Get diagonal plaquettes with custom schedules.
-        
+
         Args:
             reset: basis of the reset operation performed on data-qubits
             measurement: basis of the measurement operation performed on data-qubits
             reset_and_measured_indices: data-qubit indices that should be impacted
-            
+
         Returns:
             a mapping with 4 plaquettes: X and Z basis, both orientations
+
         """
         # _r/_m: reset/measurement basis applied to each data-qubit
         _r = reset.value.lower() if reset is not None else "-"
         _m = measurement.value.lower() if measurement is not None else "-"
-        
+
         # rs/ms: resets/measurements basis applied for each data-qubit
         rs = [_r if i in reset_and_measured_indices else "-" for i in range(4)]
         ms = [_m if i in reset_and_measured_indices else "-" for i in range(4)]
-        
+
         # Diagonal schedules:
         # X: [1, 4, 3, 2] (instead of original [1, 2, 3, 5])
         # Z: [6, 4, 3, 5] (instead of original [1, 2, 3, 5])
@@ -107,12 +116,12 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
                 ),
             },
         }
-    
+
     def get_diagonal_2_body_rpng_descriptions(
         self,
     ) -> dict[Basis, dict[PlaquetteOrientation, RPNGDescription]]:
         """Get diagonal 2-body boundary plaquettes.
-        
+
         These are derived from the diagonal bulk plaquettes by omitting qubits.
         """
         return {
@@ -131,7 +140,7 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
                 PlaquetteOrientation.RIGHT: RPNGDescription.from_string("-z6- ---- -z3- ----"),
             },
         }
-    
+
     def get_2_body_rpng_descriptions(
         self,
         reset: Basis | None = None,
@@ -139,7 +148,7 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
     ) -> dict[Basis, dict[PlaquetteOrientation, RPNGDescription]]:
         """Get 2-body boundary plaquettes (wrapper for diagonal version)."""
         return self.get_diagonal_2_body_rpng_descriptions()
-    
+
     def get_bulk_rpng_descriptions(
         self,
         reset: Basis | None = None,
@@ -147,25 +156,27 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
         reset_and_measured_indices: tuple[Literal[0, 1, 2, 3], ...] = (0, 1, 2, 3),
     ) -> dict[Basis, dict[Orientation, RPNGDescription]]:
         """Get bulk plaquettes (wrapper for diagonal version)."""
-        return self.get_diagonal_bulk_rpng_descriptions(reset, measurement, reset_and_measured_indices)
-    
+        return self.get_diagonal_bulk_rpng_descriptions(
+            reset, measurement, reset_and_measured_indices
+        )
+
     def get_3_body_rpng_descriptions(
         self,
         reset: Basis | None = None,
         measurement: Basis | None = None,
     ) -> tuple[RPNGDescription, RPNGDescription, RPNGDescription, RPNGDescription]:
         """Return the four 3-body stabilizer measurement plaquettes for diagonal schedule.
-        
+
         These correspond to corner plaquettes where two adjacent arms are missing.
         Each connects to 3 data qubits (omitting one qubit, indicated by ----).
-        
+
         For diagonal schedule:
         - Z plaquettes use schedule [6,4,3,5] instead of [1,4,3,5]
         - X plaquettes use schedule [1,4,3,2] instead of [1,2,3,5]
         """
         r = reset.value.lower() if reset is not None else "-"
         m = measurement.value.lower() if measurement is not None else "-"
-        
+
         # Diagonal 3-body plaquettes
         return (
             RPNGDescription.from_string(f"---- {r}z4{m} {r}z3{m} {r}z5{m}"),  # Top-left Z
@@ -173,7 +184,7 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             RPNGDescription.from_string(f"{r}x1{m} {r}x4{m} ---- {r}x2{m}"),  # Bottom-left X
             RPNGDescription.from_string(f"{r}z6{m} {r}z4{m} {r}z3{m} ----"),  # Bottom-right Z
         )
-    
+
     def get_memory_qubit_rpng_descriptions(
         self,
         z_orientation: Orientation = Orientation.HORIZONTAL,
@@ -191,11 +202,11 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
         # Hook errors orientations
         zhook = z_orientation.flip()
         xhook = zhook.flip()
-        
+
         # Get diagonal plaquette descriptions
         bulk_descriptions = self.get_diagonal_bulk_rpng_descriptions(reset, measurement)
         two_body_descriptions = self.get_diagonal_2_body_rpng_descriptions()
-        
+
         # Return a FrozenDefaultDict like the original
         return FrozenDefaultDict(
             {
@@ -209,7 +220,7 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             },
             default_value=RPNGDescription.empty(),
         )
-    
+
     def get_spatial_cube_qubit_rpng_descriptions(
         self,
         spatial_boundary_basis: Basis,
@@ -218,28 +229,24 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
         measurement: Basis | None = None,
     ) -> FrozenDefaultDict[int, RPNGDescription]:
         """Return RPNG descriptions for spatial cube qubit plaquettes with diagonal schedule."""
-        from tqec.utils.exceptions import TQECError
-        
         # Check if arms is invalid
         if len(arms) == 0 or len(arms) == 1:
-            raise TQECError(
-                f"Spatial cube must have at least 2 arms. Got {arms}."
-            )
+            raise TQECError(f"Spatial cube must have at least 2 arms. Got {arms}.")
         if arms in SpatialArms.I_shaped_arms():
             raise TQECError(
                 "I-shaped spatial junctions should not use get_spatial_cube_qubit_template."
             )
-        
+
         # Get parity information
         boundary_is_z = spatial_boundary_basis == Basis.Z
-        
+
         # Pre-define collection of plaquette descriptions
         corner_descriptions = self.get_3_body_rpng_descriptions(reset, measurement)
         bulk_descriptions = self.get_diagonal_bulk_rpng_descriptions(reset, measurement)
         two_body_descriptions = self.get_diagonal_2_body_rpng_descriptions()
-        
+
         mapping: dict[int, RPNGDescription] = {}
-        
+
         ####################
         #    Boundaries    #
         ####################
@@ -250,14 +257,16 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             mapping[corner] = mapping[bulk] = two_body_descriptions[_sbb][PlaquetteOrientation.UP]
         if SpatialArms.RIGHT not in arms:
             corner, bulk = (4, 21) if boundary_is_z else (2, 22)
-            mapping[corner] = mapping[bulk] = two_body_descriptions[_sbb][PlaquetteOrientation.RIGHT]
+            mapping[corner] = mapping[bulk] = two_body_descriptions[_sbb][
+                PlaquetteOrientation.RIGHT
+            ]
         if SpatialArms.DOWN not in arms:
             corner, bulk = (4, 23) if boundary_is_z else (3, 24)
             mapping[corner] = mapping[bulk] = two_body_descriptions[_sbb][PlaquetteOrientation.DOWN]
         if SpatialArms.LEFT not in arms:
             corner, bulk = (1, 12) if boundary_is_z else (3, 11)
             mapping[corner] = mapping[bulk] = two_body_descriptions[_sbb][PlaquetteOrientation.LEFT]
-        
+
         # Remove corner if both adjacent arms missing
         if SpatialArms.LEFT not in arms and SpatialArms.UP not in arms and boundary_is_z:
             del mapping[1]
@@ -267,34 +276,34 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             del mapping[3]
         if SpatialArms.RIGHT not in arms and SpatialArms.DOWN not in arms and boundary_is_z:
             del mapping[4]
-        
+
         ####################
         #       Bulk       #
         ####################
         # Hook orientations based on arms
         zup = zdown = Orientation.VERTICAL if boundary_is_z else Orientation.HORIZONTAL
         zright = zleft = zup.flip()
-        
+
         # Flip hook error orientation if arm is missing
         zup = zup if SpatialArms.UP in arms else zup.flip()
         zdown = zdown if SpatialArms.DOWN in arms else zdown.flip()
         zright = zright if SpatialArms.RIGHT in arms else zright.flip()
         zleft = zleft if SpatialArms.LEFT in arms else zleft.flip()
-        
+
         xup, xdown, xright, xleft = (zup.flip(), zdown.flip(), zright.flip(), zleft.flip())
-        
+
         # Set bulk Z plaquettes
         mapping[5] = mapping[13] = bulk_descriptions[Basis.Z][zup]
         mapping[8] = mapping[15] = bulk_descriptions[Basis.Z][zdown]
         mapping[14] = bulk_descriptions[Basis.Z][zright]
         mapping[16] = bulk_descriptions[Basis.Z][zleft]
-        
+
         # Set bulk X plaquettes
         mapping[6] = mapping[17] = bulk_descriptions[Basis.X][xup]
         mapping[7] = mapping[19] = bulk_descriptions[Basis.X][xdown]
         mapping[18] = bulk_descriptions[Basis.X][xright]
         mapping[20] = bulk_descriptions[Basis.X][xleft]
-        
+
         # Override corner plaquettes to 3-body when both adjacent arms missing
         if SpatialArms.LEFT not in arms and SpatialArms.UP not in arms and boundary_is_z:
             mapping[5] = corner_descriptions[0]
@@ -304,7 +313,7 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             mapping[7] = corner_descriptions[2]
         if SpatialArms.RIGHT not in arms and SpatialArms.DOWN not in arms and boundary_is_z:
             mapping[8] = corner_descriptions[3]
-        
+
         # Sanity check
         bulk_plaquette_indices = set(range(5, 9)) | set(range(13, 21))
         missing_bulk_plaquette_indices = bulk_plaquette_indices - mapping.keys()
@@ -312,9 +321,9 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             "Some plaquette(s) in the bulk were not correctly assigned to a "
             f"RPNGDescription. Missing indices: {missing_bulk_plaquette_indices}."
         )
-        
+
         return FrozenDefaultDict(mapping, default_value=RPNGDescription.empty())
-    
+
     def get_spatial_cube_arm_rpng_descriptions(
         self,
         spatial_boundary_basis: Basis,
@@ -324,15 +333,13 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
         measurement: Basis | None = None,
     ) -> FrozenDefaultDict[int, RPNGDescription]:
         """Return RPNG descriptions for spatial cube arm plaquettes with diagonal schedule."""
-        from tqec.utils.exceptions import TQECError
-        
         if len(arms) == 2 and arms not in SpatialArms.I_shaped_arms():
             raise TQECError(
                 f"The two provided arms cannot form a spatial pipe. Got {arms} but "
                 f"expected either a single {SpatialArms.__name__} or two but in a "
                 f"line (e.g., {SpatialArms.I_shaped_arms()})."
             )
-        
+
         if arms in [SpatialArms.LEFT, SpatialArms.RIGHT, SpatialArms.LEFT | SpatialArms.RIGHT]:
             return self._get_left_right_spatial_cube_arm_rpng_descriptions(
                 spatial_boundary_basis, arms, linked_cubes, reset, measurement
@@ -342,7 +349,7 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
                 spatial_boundary_basis, arms, linked_cubes, reset, measurement
             )
         raise TQECError(f"Got an invalid arm: {arms}.")
-    
+
     def _get_left_right_spatial_cube_arm_rpng_descriptions(
         self,
         spatial_boundary_basis: Basis,
@@ -353,18 +360,24 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
     ) -> FrozenDefaultDict[int, RPNGDescription]:
         """Left-right spatial cube arm RPNG descriptions with diagonal schedule."""
         # Bulk plaquettes for left and right parts with different reset/measure indices
-        left_boundary_descriptions = self.get_diagonal_bulk_rpng_descriptions(reset, measurement, (1, 3))
-        right_boundary_descriptions = self.get_diagonal_bulk_rpng_descriptions(reset, measurement, (0, 2))
+        left_boundary_descriptions = self.get_diagonal_bulk_rpng_descriptions(
+            reset, measurement, (1, 3)
+        )
+        right_boundary_descriptions = self.get_diagonal_bulk_rpng_descriptions(
+            reset, measurement, (0, 2)
+        )
         two_body_descriptions = self.get_diagonal_2_body_rpng_descriptions()
-        
+
         # Hook errors adapted to boundary basis
-        zhook = Orientation.HORIZONTAL if spatial_boundary_basis == Basis.Z else Orientation.VERTICAL
+        zhook = (
+            Orientation.HORIZONTAL if spatial_boundary_basis == Basis.Z else Orientation.VERTICAL
+        )
         xhook = zhook.flip()
-        
+
         # Plaquette indices
         up = 2 if spatial_boundary_basis == Basis.Z else 1
         down = 3 if spatial_boundary_basis == Basis.Z else 4
-        
+
         mapping = {
             up: two_body_descriptions[spatial_boundary_basis][PlaquetteOrientation.UP],
             down: two_body_descriptions[spatial_boundary_basis][PlaquetteOrientation.DOWN],
@@ -373,12 +386,12 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             7: right_boundary_descriptions[Basis.X][xhook],
             8: right_boundary_descriptions[Basis.Z][zhook],
         }
-        
+
         # Handle 3-body corner plaquettes
         _corners = self.get_3_body_rpng_descriptions(reset, measurement)
         u, v = linked_cubes
         _sbb = spatial_boundary_basis
-        
+
         # Replace top plaquette if it should be a 3-body stabilizer
         if SpatialArms.LEFT in arms and _sbb == Basis.Z and SpatialArms.UP in v.spatial_arms:
             mapping[up] = _corners[0]
@@ -389,9 +402,9 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             mapping[down] = _corners[2]
         if SpatialArms.RIGHT in arms and _sbb == Basis.Z and SpatialArms.DOWN in u.spatial_arms:
             mapping[down] = _corners[3]
-        
+
         return FrozenDefaultDict(mapping, default_value=RPNGDescription.empty())
-    
+
     def _get_up_down_spatial_cube_arm_rpng_descriptions(
         self,
         spatial_boundary_basis: Basis,
@@ -403,17 +416,21 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
         """Up-down spatial cube arm RPNG descriptions with diagonal schedule."""
         # Bulk plaquettes for up and down parts
         up_bulk_descriptions = self.get_diagonal_bulk_rpng_descriptions(reset, measurement, (2, 3))
-        down_bulk_descriptions = self.get_diagonal_bulk_rpng_descriptions(reset, measurement, (0, 1))
+        down_bulk_descriptions = self.get_diagonal_bulk_rpng_descriptions(
+            reset, measurement, (0, 1)
+        )
         two_body_descriptions = self.get_diagonal_2_body_rpng_descriptions()
-        
+
         # Hook errors adapted to boundary basis
-        zhook = Orientation.VERTICAL if spatial_boundary_basis == Basis.Z else Orientation.HORIZONTAL
+        zhook = (
+            Orientation.VERTICAL if spatial_boundary_basis == Basis.Z else Orientation.HORIZONTAL
+        )
         xhook = zhook.flip()
-        
+
         # Plaquette indices
         left = 3 if spatial_boundary_basis == Basis.Z else 1
         right = 2 if spatial_boundary_basis == Basis.Z else 4
-        
+
         mapping = {
             left: two_body_descriptions[spatial_boundary_basis][PlaquetteOrientation.LEFT],
             right: two_body_descriptions[spatial_boundary_basis][PlaquetteOrientation.RIGHT],
@@ -422,12 +439,12 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             7: down_bulk_descriptions[Basis.X][xhook],
             8: down_bulk_descriptions[Basis.Z][zhook],
         }
-        
+
         # Handle 3-body corner plaquettes
         _corners = self.get_3_body_rpng_descriptions(reset, measurement)
         u, v = linked_cubes
         _sbb = spatial_boundary_basis
-        
+
         # Replace left plaquette if it should be a 3-body stabilizer
         if SpatialArms.UP in arms and _sbb == Basis.Z and SpatialArms.LEFT in v.spatial_arms:
             mapping[left] = _corners[0]
@@ -438,24 +455,19 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             mapping[right] = _corners[1]
         if SpatialArms.DOWN in arms and _sbb == Basis.Z and SpatialArms.RIGHT in u.spatial_arms:
             mapping[right] = _corners[3]
-        
+
         return FrozenDefaultDict(mapping, default_value=RPNGDescription.empty())
-    
+
     def get_memory_qubit_raw_template(self):
-        """Return the template instance needed to implement a standard memory operation on a logical qubit."""
-        from tqec.templates.qubit import QubitTemplate
+        """Return the template for a standard memory operation on one logical qubit."""
         return QubitTemplate()
-    
+
     def get_spatial_cube_qubit_raw_template(self):
         """Return the template instance needed to implement a spatial cube qubit."""
-        from tqec.templates.qubit import QubitSpatialCubeTemplate
         return QubitSpatialCubeTemplate()
-    
+
     def get_spatial_cube_arm_raw_template(self, arms: SpatialArms):
         """Return the template instance needed to implement the given spatial arms."""
-        from tqec.utils.exceptions import TQECError
-        from tqec.templates.qubit import QubitVerticalBorders, QubitHorizontalBorders
-        
         if (
             len(arms) == 0
             or len(arms) > 2
@@ -472,18 +484,18 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
             return QubitHorizontalBorders()
         else:
             raise TQECError(f"Unrecognized spatial arm(s): {arms}.")
-    
+
     def get_memory_qubit_plaquettes(
         self,
         z_orientation: Orientation = Orientation.HORIZONTAL,
         reset: Basis | None = None,
         measurement: Basis | None = None,
     ):
-        """Return the plaquettes needed to implement a standard memory operation on a logical qubit."""
+        """Return the plaquettes needed for a standard memory operation."""
         return self._mapper(self.get_memory_qubit_rpng_descriptions)(
             z_orientation, reset, measurement
         )
-    
+
     def get_spatial_cube_qubit_plaquettes(
         self,
         spatial_boundary_basis: Basis,
@@ -495,7 +507,7 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
         return self._mapper(self.get_spatial_cube_qubit_rpng_descriptions)(
             spatial_boundary_basis, arms, reset, measurement
         )
-    
+
     def get_spatial_cube_arm_plaquettes(
         self,
         spatial_boundary_basis: Basis,
@@ -508,4 +520,3 @@ class DiagonalScheduleGenerator(FixedBulkConventionGenerator):
         return self._mapper(self.get_spatial_cube_arm_rpng_descriptions)(
             spatial_boundary_basis, arms, linked_cubes, reset, measurement
         )
-
