+++ src/tqec/plaquette/rpng/translators/diagonal.py
@@ -1,0 +1,7 @@
+from tqec.plaquette.rpng.translators.scheduled import ScheduledRPNGTranslator
+
+
+class DiagonalRPNGTranslator(ScheduledRPNGTranslator):
+    """Translator for diagonal-schedule plaquettes."""
+
+    MEASUREMENT_SCHEDULE = 7
